refactor(deployment): break domain→app circular import in cancellation flow
