+++ apps/backend/src/modules/deployment/deployment.domain.ts
@@ -6,7 +6,6 @@ import {
   DeploymentRequestPlatformState,
   PlatformIdentifier,
   QueryDeploymentRequestsListArgs,
-  ServiceInstanceCreationStatus,
 } from '../../__generated__/resolvers-types';
 import { withTransaction } from '../../context/database.context';
 import DeploymentRequest, {
@@ -15,29 +14,15 @@ import DeploymentRequest, {
   DeploymentRequestMutator,
 } from '../../model/kanel/public/DeploymentRequest';
 import { OrganizationId } from '../../model/kanel/public/Organization';
-import { UserId } from '../../model/kanel/public/User';
 import { auth0Client } from '../../thirdparty/auth0/client';
 import { logApp } from '../../utils/app-logger.util';
-import {
-  ErrorCode,
-  NotFoundErrorCode,
-  UnknownErrorCode,
-} from '../../utils/error/error.code';
+import { ErrorCode, UnknownErrorCode } from '../../utils/error/error.code';
 import { prefixObjectKeys } from '../../utils/utils';
-import { ServiceInstanceDomain } from '../service/instance/service-instance.domain';
-import {
-  BUNDLE_REQUEST_CANCELLATION_REASON,
-  DeploymentApp,
-} from './deployment.app';
 import {
   ServiceGroupDomain,
   ServiceGroupName,
 } from './group/service-group.domain';
-import {
-  DeploymentQuotaDomain,
-  QuotaKey,
-  quotaKeysOfRequest,
-} from './quota/deployment.quota.domain';
+import { QuotaKey } from './quota/deployment.quota.domain';
 
 const scopeToPlatformIdentifier =
   (key: QuotaKey) => (builder: Knex.QueryBuilder<DeploymentRequest>) => {
@@ -472,71 +457,6 @@ export const DeploymentRequestDomain = {
     });
   },
 
-  applyCancellationToDeploymentRequest: async ({
-    deploymentRequest,
-    userId,
-    isAdmin,
-    cancellationReason,
-  }: {
-    deploymentRequest: DeploymentRequest;
-    userId: UserId;
-    isAdmin: boolean;
-    cancellationReason?: string;
-  }): Promise<DeploymentRequest> => {
-    const previousHubStatus = deploymentRequest.hub_status;
-
-    const countsInOrgaQuota =
-      isAdmin ||
-      ![
-        DeploymentRequestPlatformState.Unprovisioned,
-        DeploymentRequestPlatformState.Provisioning,
-      ].includes(
-        deploymentRequest.actual_state ??
-          DeploymentRequestPlatformState.Unprovisioned
-      );
-
-    const target_state =
-      deploymentRequest.actual_state ===
-      DeploymentRequestPlatformState.Unprovisioned
-        ? DeploymentRequestPlatformState.Unprovisioned
-        : DeploymentRequestPlatformState.Removed;
-
-    return DeploymentQuotaDomain.withLockedQuotaTransaction(
-      quotaKeysOfRequest(deploymentRequest),
-      async () => {
-        const updatedDeploymentRequest =
-          await DeploymentRequestDomain.updateDeploymentRequestById(
-            deploymentRequest.id,
-            {
-              hub_status: DeploymentRequestHubStatus.Cancelled,
-              target_state: target_state,
-              cancellation_date: new Date(),
-              cancellation_user_id: userId,
-              cancellation_reason: cancellationReason,
-              counts_in_orga_quota: countsInOrgaQuota,
-            }
-          );
-        if (!updatedDeploymentRequest) {
-          throw new Error(NotFoundErrorCode.DeploymentRequestNotFound);
-        }
-        if (!countsInOrgaQuota) {
-          await ServiceInstanceDomain.updateServiceInstance(
-            deploymentRequest.service_instance_id,
-            {
-              creation_status: ServiceInstanceCreationStatus.Disabled,
-            }
-          );
-        }
-        await DeploymentApp.releaseDeploymentRequestPlace(
-          previousHubStatus,
-          deploymentRequest
-        );
-
-        return updatedDeploymentRequest;
-      }
-    );
-  },
-
   deleteDeploymentRequestAudience: async (
     deploymentRequest: DeploymentRequest
   ): Promise<void> => {
@@ -561,25 +481,6 @@ export const DeploymentRequestDomain = {
         error,
         deploymentRequestId: deploymentRequest.id,
       });
-    }
-  },
-
-  cancelOngoingStandaloneTrialsForBundle: async (
-    bundle: DeploymentRequest
-  ): Promise<void> => {
-    const standaloneTrials =
-      await DeploymentRequestDomain.loadOngoingStandaloneTrialsForOrganization(
-        bundle.organization_requester_id
-      );
-
-    for (const trial of standaloneTrials) {
-      await DeploymentRequestDomain.applyCancellationToDeploymentRequest({
-        deploymentRequest: trial,
-        userId: bundle.user_requester_id,
-        isAdmin: false,
-        cancellationReason: BUNDLE_REQUEST_CANCELLATION_REASON,
-      });
-      await DeploymentRequestDomain.deleteDeploymentRequestAudience(trial);
     }
   },
 };
